refactor: extract common functionality
