+++ src/v3/ConsoleTraceLogger.ts
@@ -1,15 +1,16 @@
-import {
-  DEFAULT_DEBOUNCE_DURATION,
-  DEFAULT_INTERACTIVE_TIMEOUT_DURATION,
-} from './constants'
 import type {
   AllPossibleRequiredSpanSeenEvents,
   AllPossibleStateTransitionEvents,
   AllPossibleTraceStartEvents,
 } from './debugTypes'
-import { isSuppressedError } from './debugUtils'
+import {
+  extractTimingOffsets,
+  formatMatcher,
+  getComputedResults,
+  getConfigSummary,
+  isSuppressedError,
+} from './debugUtils'
 import type { SpanMatcherFn } from './matchSpan'
-import { createTraceRecording } from './recordingComputeUtils'
 import type { FinalTransition, OnEnterStatePayload } from './Trace'
 import { isTerminalState } from './Trace'
 import type { TraceManager } from './TraceManager'
@@ -32,73 +33,6 @@ interface ConsoleLike {
   groupCollapsed: (...args: any[]) => void
   groupEnd: () => void
   // Add other console methods if needed (warn, error, etc.)
-}
-
-/**
- * Formats a SpanMatcherFn into a more readable string representation.
- * This is a basic implementation and can be significantly improved
- * based on the actual structure of `matcher.fromDefinition`.
- *
- * @param matcher The matcher function to format.
- * @param index Optional index for generic naming.
- * @returns A string representation of the matcher.
- */
-export function formatMatcher<
-  SelectedRelationNameT extends keyof RelationSchemasT,
-  RelationSchemasT extends RelationSchemasBase<RelationSchemasT>,
-  VariantsT extends string,
->(
-  matcher: SpanMatcherFn<SelectedRelationNameT, RelationSchemasT, VariantsT>,
-  index?: number,
-): string {
-  // Check if the matcher has attached definition info
-  if (matcher.fromDefinition) {
-    // Attempt to create a more descriptive name from the definition
-    // This part needs customization based on how 'fromDefinition' is structured.
-    // Example: Check for specific properties like 'name', 'type', 'label' etc.
-    const def = matcher.fromDefinition
-    const parts: string[] = []
-
-    if (typeof def === 'object' && def !== null) {
-      // Example: Prioritize 'label' if it exists
-      if ('label' in def && typeof def.label === 'string') {
-        return `Label: "${def.label}"`
-      }
-      // Example: Use 'name' if it exists
-      if ('name' in def) {
-        if (typeof def.name === 'string') {
-          parts.push(`Name: "${def.name}"`)
-        } else if (def.name instanceof RegExp) {
-          parts.push(`Name: /${def.name.source}/${def.name.flags}`)
-        }
-      }
-      // Example: Add type if present
-      if ('type' in def && typeof def.type === 'string') {
-        parts.push(`Type: ${def.type}`)
-      }
-      // Add other relevant properties from your definition structure
-    }
-
-    if (parts.length > 0) {
-      return parts.join(', ')
-    }
-
-    // Fallback: Stringify the definition (can be verbose)
-    try {
-      const defString = JSON.stringify(def)
-      // Limit length to avoid overly long strings
-      return defString.length > 100
-        ? // eslint-disable-next-line no-magic-numbers
-          `${defString.slice(0, 97)}...`
-        : defString
-    } catch {
-      // Fallback if stringify fails
-      return `Matcher Definition #${index ?? '?'}`
-    }
-  }
-
-  // Fallback if no definition info is available
-  return `Matcher #${index ?? '?'}`
 }
 
 /**
@@ -154,45 +88,6 @@ interface TraceInfo<RelationSchemasT> {
   attributes?: Record<string, unknown>
   relatedTo?: Record<string, unknown>
   requiredSpans: { name: string; isMatched: boolean; matcher: Function }[]
-}
-
-/**
- * Extract timing offsets from a transition object
- */
-const extractTimingOffsets = <RelationSchemasT>(
-  transition: OnEnterStatePayload<RelationSchemasT>,
-) => {
-  // ... (implementation unchanged) ...
-  let lastRequiredSpanOffset: number | undefined
-  let completeSpanOffset: number | undefined
-  let cpuIdleSpanOffset: number | undefined
-
-  if (
-    'lastRequiredSpanAndAnnotation' in transition &&
-    transition.lastRequiredSpanAndAnnotation
-  ) {
-    lastRequiredSpanOffset =
-      transition.lastRequiredSpanAndAnnotation.annotation
-        .operationRelativeEndTime
-  }
-
-  if (
-    'completeSpanAndAnnotation' in transition &&
-    transition.completeSpanAndAnnotation
-  ) {
-    completeSpanOffset =
-      transition.completeSpanAndAnnotation.annotation.operationRelativeEndTime
-  }
-
-  if (
-    'cpuIdleSpanAndAnnotation' in transition &&
-    transition.cpuIdleSpanAndAnnotation
-  ) {
-    cpuIdleSpanOffset =
-      transition.cpuIdleSpanAndAnnotation.annotation.operationRelativeEndTime
-  }
-
-  return { lastRequiredSpanOffset, completeSpanOffset, cpuIdleSpanOffset }
 }
 
 /**
@@ -474,16 +369,7 @@ export function createConsoleTraceLogger<
     }
     log(`   Required spans: ${requiredSpans.length}`)
     // Log config summary
-    const variantConfig = trace.definition.variants[trace.input.variant]
-    const timeout = variantConfig?.timeout
-    const debounce =
-      trace.definition.debounceWindow ?? DEFAULT_DEBOUNCE_DURATION
-    const interactive =
-      typeof trace.definition.captureInteractive === 'object'
-        ? trace.definition.captureInteractive.timeout
-        : trace.definition.captureInteractive
-        ? DEFAULT_INTERACTIVE_TIMEOUT_DURATION
-        : undefined
+    const { timeout, debounce, interactive } = getConfigSummary(trace)
     log(
       `   Config: Timeout=${timeout}ms, Debounce=${debounce}ms${
         interactive ? `, Interactive=${interactive}ms` : ''
@@ -550,12 +436,13 @@ export function createConsoleTraceLogger<
         logTimingInfo(transition)
         // Log computed values and spans
         try {
-          const rec = createTraceRecording(trace, transition)
-          const computedVals = rec.computedValues
-          const { computedSpans } = rec
-          if (computedVals && Object.keys(computedVals).length > 0) {
+          const { computedValues, computedSpans } = getComputedResults(
+            trace,
+            transition,
+          )
+          if (computedValues && Object.keys(computedValues).length > 0) {
             log(
-              `   Computed Values: ${Object.entries(computedVals)
+              `   Computed Values: ${Object.entries(computedValues)
                 .map(([k, v]) => `${k}=${v}`)
                 .join(', ')}`,
             )
@@ -587,15 +474,13 @@ export function createConsoleTraceLogger<
         logTimingInfo(transition)
         // Log computed values and spans
         try {
-          const rec = createTraceRecording(
+          const { computedValues, computedSpans } = getComputedResults(
             trace,
             transition as FinalTransition<RelationSchemasT>,
           )
-          const computedVals = rec.computedValues
-          const { computedSpans } = rec
-          if (computedVals && Object.keys(computedVals).length > 0) {
+          if (computedValues && Object.keys(computedValues).length > 0) {
             log(
-              `   Computed Values: ${Object.entries(computedVals)
+              `   Computed Values: ${Object.entries(computedValues)
                 .map(([k, v]) => `${k}=${v}`)
                 .join(', ')}`,
             )
